+++ ClientApp/src/app/models/requests/photoToUpload.ts
