+++ ClientApp/src/app/models/requests/saveProduct.ts
