+++ ClientApp/src/app/models/requests/saveSupplyEntry.ts
