+++ ClientApp/src/app/models/responses/photo.ts
