+++ ClientApp/src/app/models/responses/product.ts
