+++ ClientApp/src/app/models/responses/productPhoto.ts
