+++ ClientApp/src/app/models/responses/supply.ts
