+++ ClientApp/src/app/models/responses/technician.ts
